Refactor view models and namespaces for optimization features
