+++ Tunny/WPF/Views/Pages/HelpPage.xaml
@@ -4,14 +4,10 @@
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-    xmlns:viewModels="clr-namespace:Tunny.WPF.ViewModels"
     Title="HelpPage"
     d:DesignHeight="600"
     d:DesignWidth="1000"
     mc:Ignorable="d">
-    <Page.DataContext>
-        <viewModels:OptimizeViewModel />
-    </Page.DataContext>
 
     <Grid Margin="10">
         <Grid.RowDefinitions>
