+++ Tunny/WPF/Views/Pages/Optimize/OptimizePage.xaml
@@ -5,7 +5,7 @@
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
     xmlns:lvc="clr-namespace:LiveChartsCore.SkiaSharpView.WPF;assembly=LiveChartsCore.SkiaSharpView.WPF"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-    xmlns:viewModels="clr-namespace:Tunny.WPF.ViewModels"
+    xmlns:viewModels="clr-namespace:Tunny.WPF.ViewModels.Optimize"
     Title="OptimizePage"
     d:DesignHeight="650"
     d:DesignWidth="1000"
